refactor(ui): extract product config panel
